Add `AWS-SDK` V2 and keep backward compatibility with V1 plugins
If you include the module `LogStash::PluginMixins::AwsConfig` it will
use the V1 Api, if your plugin require the V2 Api make sure to include
the `LogStash::PluginMixins::AwsConfig::V2`.

This change was necessary because of how our dependencies works, if you
install a plugin that need the V2 api it will install a newer version of the
mixin. Previously installed plugins dependeing on the v1 api wont be updated.

Changes include:
- The V2 module also use the `Aws::Credentials` class when generating
  the options hash.
- If the plugin doesn't implement the `#aws_service_endpoint` method it
  will fallback to `{ :region => 'east-1'}` (V2 only)
- `use_ssl` option is now deprecated and its not in use in the V2 code, the
  new version of the API now set it to true and only recommended to turn
  it off when doing local testing.

fixes #8

Fixes #13